resolved build issues
alltough the software is not working, the messages and services are compiled which are the deciding factor on this machine

+++ src/map.cpp
@@ -165,7 +165,7 @@ class Map : public rclcpp::Node
 
             if (value >= -128 && value <= 127)
             {
-                RCLCPP_INFO(get_logger(), "c: %d", value);
+                RCLCPP_INFO(get_logger(), "c: %d", (int)value);
                 if (value != GridColor::Grey)
                 {
                     static_map_colors.push_back(GridColor::Black);
